+++ CHANGELOG.md
@@ -15,6 +15,8 @@ Releases marked with 🧪 (or previously with the "beta" suffix) were released o
 * 🔨 Movies: when connecting Cloud or Trakt, movies might not have been added if downloading them
   failed. Any missing movies will be added on the next sync.
 * 🔨 Movies: prefer translated title when viewing details.
+* 🔧 Movies: at least display a more helpful error message when updating a movie fails because it
+  was removed from TMDB.
 * 📝 Devices running Android 5.0 and 5.1 will receive no more updates in the future.
 * 📝 Import latest user interface translations.