+++ CHANGELOG.md
@@ -21,6 +21,8 @@ Releases marked with 🧪 (or previously with the "beta" suffix) were released o
   make removing the secondary action.
 * 🔧 Shows: display indicator on tab icon if a general or a stream and purchase filter is active.
 * 🔧 Widget: update preview to include more items, remove old settings button.
+* 🔧 Update message when there is no description to reflect that there can also be no details at
+  all, not just a missing translation. 
 
 ### 2025.2.0 - 2025-03-27 🧪